PUBDEV-6171: based on the test, I used a different dataset that will run much faster.
use smaller dataset.
